#4421 #4420

+++ extensions/innertube/src/main/kotlin/me/knighthat/common/response/AudioFormat.kt
@@ -2,11 +2,11 @@ package me.knighthat.common.response
 
 interface AudioFormat: Comparable<AudioFormat> {
 
-    val itag: UByte
+    val itag: UShort
     val url: String
     val mimeType: String
     val codec: String
-    val bitrate: Int
+    val bitrate: UInt
 
     /**
      * Using bitrate as the value to compare between objects
